+++ src/Order/CartComposer.php
@@ -61,7 +61,7 @@ class CartComposer
         }
         // load desktop cart, duplicate its context, add info to context & create new cart based on it
         $initContext = $this->contextComposer->getContextByCustomerId($customerId ?? '');
-        $cleanCartContext = $this->contextManager->duplicateContextWithNewToken($initContext, $customerId);
+        $cleanCartContext = $this->contextManager->duplicateContextWithNewToken($initContext, $customerId ?? '');
         $this->contextComposer->addCustomerAddress($sgCart, $cleanCartContext);
         $paymentId = $this->paymentComposer->mapIncomingPayment($sgCart, $cleanCartContext);
         $context = $this->contextComposer->addActivePayment($paymentId, $cleanCartContext);