+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Pre-version
 
+- added `set_settings`
 - fixed customer address type mapping in case billing and shipping address are the same
 
 # 1.6.4