+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - added `get_reviews`
 - added Shopgate `custom_fields` support to customer & customer addresses (see Readme)
 - fixed customer address type mapping in case billing and shipping address are the same
+- fixed product sort to use parent ID instead of child ID for sort order positions in a category
 
 # 1.6.4