+++ CHANGELOG.md
@@ -2,8 +2,9 @@
 
 - added `set_settings`
 - added `get_reviews`
-- added Shopgate `custom_fields` support to customer & customer addresses (see Readme)
-- added support for categories with dynamic product groups
+- added Shopgate `custom_fields` support to customer & customer addresses (see [README](README.md))
+- added support for categories with `dynamic product groups`
+- added support for product `CrossSell` sliders (max 4)
 - fixed customer address type mapping in case billing and shipping address are the same
 - fixed product sort to use parent ID instead of child ID for sort order positions in a category